// app/src/Data/Types/Creature.ts
export default interface Creature {
  Type: string;
  Rank: number;
  Health: number;
  Classification: string;
  School: School;
  Masteries: School[],
  Cheats: boolean,
  StartPips: number;
  PowerPips: boolean;
  OutPierce?: string;
  OutBoost?: string;
  IncBoost?: string;
  IncResist?: string;
  OutHealing: string;
  IncHealing: string;
  ShadowPipSlots: number;
  Stunable: boolean;
  Beguilable: boolean;
  Minion?: string;
  Minion2?: string;
  World: string;
  Location: string;
  Sublocation?: string;
  Description: string;
  Speech?: string;
  MonstrotomeDescription: string;
  SummonAnimus: number;
  SummonGold: number;
  GuestAnimus: number;
  GuestGold: number;
  CanExpel: boolean;
  ExpelAnimus: number;
  ExpelGold: number;
  SpellNotes?: string;
  Casts: string[];
  GoldRange: [number, number];
  Drops: {
    Hats: string[];
    Robes: string[];
    Boots: string[];
    Athames: string[];
    Amulets: string[];
    Rings: string[];
    Wands: string[];
    Decks: string[];
    HousingItems: string[];
    Spellements: string[];
    Reagents: string[];
    TreasureCards: string[];
    Snacks: string[];
    Jewels: string[];
    Pets: string[];
    Mounts: string[];
    Elixirs: string[];
    Recipes: string[];
    Seeds: string[];
    Spells: string[];
  };
}